requirements.txt
absl-py==1.4.0
astunparse==1.6.3
cachetools==4.2.4
certifi==2022.12.7
charset-normalizer==3.1.0
gast==0.3.3
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.3
h5py==2.10.0
idna==3.4
imageio==2.27.0
importlib-metadata==6.1.0
install==1.3.5
joblib==1.2.0
Keras-Preprocessing==1.1.2
Markdown==3.4.3
numpy==1.23.4
oauthlib==3.2.2
opencv-python==4.7.0.72
opt-einsum==3.3.0
Pillow==9.4.0
pip==23.0.1
protobuf==3.20.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.4.1
setuptools==56.0.0
six==1.16.0
tensorboard==2.2.2
tensorboard-plugin-wit==1.8.1
tensorflow==2.2.0
tensorflow-estimator==2.2.0
termcolor==2.2.0
threadpoolctl==3.1.0
urllib3==1.26.15
Werkzeug==2.0.3
wheel==0.40.0
wrapt==1.15.0
zipp==3.15.0
